product_service/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
pymongo==4.6.1
motor==3.3.2
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.2.1
pydantic-extra-types==2.1.0
beanie==1.23.0
httpx==0.26.0

# Для разработки и тестирования
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Дополнительные утилиты
loguru==0.7.2
python-dateutil==2.8.2
typing-extensions==4.9.0